`crypto-refresh`: add support for Argon2 S2K (#5)
Added support for encrypting and decrypting messages or keys that use Argon2 as
S2K.
This feature is backwards compatible, no breaking changes, as long as argon2 is
not used.
If it is used, since Wasm is dynamically imported and used under the hood,
apps might need bundling or configuration changes to be able to handle the
module setup.

Newly introduced config fields (relevant for encryption only):
- `config.s2kType` (defaulting to `enums.s2k.iterated`): s2k to use on
password-based encryption as well as private key encryption;
- `config.s2kArgon2Params` (defaulting to "uniformly safe settings" from Argon
RFC): parameters to use on encryption when `config.s2kType` is set to
`enums.s2k.argon2`;